adding a hmi client to record pendulum angles

+++ examples/inverted_pendulum/simulation.py
@@ -75,9 +75,27 @@ def start_comm_module(path_to_plc_specification_file, ip_address_to_listen,
         print "Started Modbus comm module with pid ", newpid
         return newpid
 
+def start_example_hmi(is_virtual, rel_cpu_speed, 
+    n_insns_per_round, log_file_fd):
+    newpid = os.fork()
+    if newpid == 0:
+        os.dup2(log_file_fd, sys.stdout.fileno())
+        os.dup2(log_file_fd, sys.stderr.fileno())
+        if is_virtual == True:
+            cmd_str = "example_hmi"
+            args = ["tracer", "-c", cmd_str, "-r", str(rel_cpu_speed), "-n", \
+                str(n_insns_per_round)]
+            os.execvp(args[0], args)
+        else:
+            args = ["example_hmi"]
+            os.execvp(args[0], args)
+    else:
+        print "Started example hmi with pid ", newpid
+        return newpid
+
 
 def main(is_virtual=False,
-        num_dilated_nodes=2,
+        num_dilated_nodes=3,
         run_time=5, 
         rel_cpu_speed=1.0,
         num_insns_per_round=1000000):
@@ -109,6 +127,7 @@ def main(is_virtual=False,
     fd1 = os.open( "/tmp/pc_grpc_server_log.txt", os.O_RDWR | os.O_CREAT )
     fd2 = os.open( "/tmp/plc_log.txt", os.O_RDWR | os.O_CREAT )
     fd3 = os.open( "/tmp/comm_module_log.txt", os.O_RDWR | os.O_CREAT )
+    fd4 = os.open( "/tmp/example_hmi.txt", os.O_RDWR | os.O_CREAT )
 
     args = parser.parse_args()
      
@@ -132,6 +151,10 @@ def main(is_virtual=False,
         args.comm_module_attached_resource, is_virtual, rel_cpu_speed, \
         num_insns_per_round, fd3)
 
+    print "Starting HMI ..."
+    example_hmi_pid = start_example_hmi(is_virtual, rel_cpu_speed, \
+        num_insns_per_round, fd4)
+
 
     # Wait until all processes have started and registered themselves
     emulation.wait_for_initialization()
@@ -151,13 +174,17 @@ def main(is_virtual=False,
     os.close(fd1)
     os.close(fd2)
     os.close(fd3)
+    os.close(fd4)
 
 
 
-    print "Interrupting/Killing all spawned processes !"
+    print "Interrupting all spawned processes !"
     os.kill(grpc_server_pid, signal.SIGINT)
-    os.kill(plc_pid, signal.SIGINT)
-    os.kill(comm_module_pid, signal.SIGINT)
+    
+    if is_virtual == False:
+        os.kill(plc_pid, signal.SIGINT)
+        os.kill(comm_module_pid, signal.SIGINT)
+        os.kill(example_hmi_pid, signal.SIGINT)
 
     print "Emulation finished ! "
